refactor: delete src/server/ — inline re-exports into consumers (Plan E)
