+++ src/mcp/tools.ts
@@ -13,12 +13,16 @@ import { remoteClient } from "../activitypub/remote-client.js";
 import { dynamicInstanceDiscovery } from "../discovery/dynamic-instance-discovery.js";
 import { instanceDiscovery } from "../discovery/instance-discovery.js";
 import type { RateLimiter } from "../resilience/rate-limiter.js";
-import { validateActorIdentifier, validateDomain, validateQuery } from "../server/index.js";
 import { healthChecker } from "../telemetry/health-check.js";
 import { performanceMonitor } from "../telemetry/performance-monitor.js";
 import { formatErrorWithSuggestion, getErrorMessage } from "../utils/errors.js";
 import { stripHtmlTags } from "../utils/html.js";
 import { ActorIdentifierSchema, DomainSchema, QuerySchema } from "../validation/schemas.js";
+import {
+  validateActorIdentifier,
+  validateDomain,
+  validateQuery,
+} from "../validation/validators.js";
 import { trackedMcpServer } from "./capabilities.js";
 import { registerExportTools } from "./tools-export.js";
 import { registerWriteTools } from "./tools-write.js";
