--- src/server/index.ts
@@ -1,17 +1,0 @@
-/**
- * Server modules for the ActivityPub MCP Server
- */
-
-export {
-  AdaptiveRateLimiter,
-  adaptiveRateLimiter,
-  type InstanceRateLimit,
-} from "../resilience/adaptive-rate-limiter.js";
-export { type RateLimitConfig, RateLimiter } from "../resilience/rate-limiter.js";
-export { type HttpTransportOptions, HttpTransportServer } from "../transport/http.js";
-export {
-  extractSingleValue,
-  validateActorIdentifier,
-  validateDomain,
-  validateQuery,
-} from "../validation/validators.js";
